+++ src/main/java/com/otectus/runicskills/client/core/Utils.java
@@ -26,7 +26,11 @@ import net.minecraftforge.fml.ModList;
 
 
 public class Utils {
-    public static int FONT_COLOR = Color.WHITE.getRGB();
+    public static final int FONT_COLOR = Color.WHITE.getRGB();
+    public static final int SKILL_ABBR_COLOR = new Color(240, 240, 240).getRGB();
+    public static final int SKILL_LEVEL_COLOR = Color.WHITE.getRGB();
+    public static final int TITLE_SELECTED_COLOR = 0x55FF55;
+    public static final int TITLE_UNSELECTED_COLOR = 0xFFAA00;
     public static final Minecraft client = Minecraft.getInstance();
 
     public static ItemStack playerHead() {